scripts/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pdbstats"
version = "0.1.0"
description = "Tiny toolkit to compute per-residue/global RMSD and pLDDT statistics from predicted PDBs and plot them"
authors = [{ name = "Your Name" }]
requires-python = ">=3.9"
dependencies = [
  "biopython>=1.79",
  "pandas>=1.5",
  "numpy>=1.23",
  "matplotlib>=3.7",
  "seaborn>=0.12",
  "typer>=0.9"
]